refactor(perl-ci-hygiene): split doc path checker into module (#0000)
